requirements.txt
streamlit==1.28.0
sentence-transformers>=3.0.0
pdfplumber>=0.11.0
python-docx==0.8.11
langchain==0.1.0
langchain-community>=0.0.9
ollama==0.1.0
numpy
scikit-learn
huggingface-hub>=0.16.0
